// front/src/services/ant-design-pro/typings.d.ts
// @ts-ignore
/* eslint-disable */

import {ActionType} from "@ant-design/pro-components";
import {MutableRefObject} from "react";
import {ProlistItem} from "@/pages/live/monitor/imageList/config/type/ProlistItem";
import {LogRecord} from "@/pages/live/monitor/logRecord/config/type/logRecord"
declare namespace API {
  type UpdateFormProps = {
    actionRef?: MutableRefObject<ActionType | undefined>;
  };
  type CurrentUser = {
    name?: string;
    avatar?: string;
    userid?: string;
    email?: string;
    signature?: string;
    title?: string;
    group?: string;
    tags?: { key?: string; label?: string }[];
    notifyCount?: number;
    unreadCount?: number;
    country?: string;
    access?: string;
    roleList?: number[];
    geographic?: {
      province?: { label?: string; key?: string };
      city?: { label?: string; key?: string };
    };
    address?: string;
    phone?: string;
  };
  // 菜单
  export type Menu = Menu;
  // 角色
  export type Role = Role;
  // 用户
  export type User = User
  // 直播分类
  export type LiveCat1 = LiveCat1
  export type LiveCat2 = LiveCat2
  export type LiveCat3 = LiveCat3
  export type LogRecord = LogRecord
  // proListItem
  export type ProListItem = ProlistItem

  type LoginResult = {
    status?: string;
    type?: string;
    currentAuthority?: string;
  };

  type PageParams = {
    current?: number;
    pageSize?: number;
  };

  type RuleListItem = {
    key?: number;
    disabled?: boolean;
    href?: string;
    avatar?: string;
    name?: string;
    owner?: string;
    desc?: string;
    callNo?: number;
    status?: number;
    updatedAt?: string;
    createdAt?: string;
    progress?: number;
  };

  type RuleList = {
    data?: RuleListItem[];
    /** 列表的内容总数 */
    total?: number;
    success?: boolean;
  };

  type FakeCaptcha = {
    code?: number;
    status?: string;
  };

  type LoginParams = {
    username?: string;
    password?: string;
    autoLogin?: boolean;
    type?: string;
  };

  type ErrorResponse = {
    /** 业务约定的错误码 */
    errorCode: string;
    /** 业务上的错误信息 */
    errorMessage?: string;
    /** 业务上的请求是否成功 */
    success?: boolean;
  };

  type NoticeIconList = {
    data?: NoticeIconItem[];
    /** 列表的内容总数 */
    total?: number;
    success?: boolean;
  };

  type NoticeIconItemType = 'notification' | 'message' | 'event';

  type NoticeIconItem = {
    id?: string;
    extra?: string;
    key?: string;
    read?: boolean;
    avatar?: string;
    title?: string;
    status?: string;
    datetime?: string;
    description?: string;
    type?: NoticeIconItemType;
  };
}